Added formulas for uniques
Added pyc to gitignore, added calculation for individual probability tables as well as estimation for group points if not provided. Final response should be given as json breaking down party_prob, personal_prob, and stats used in the calculation, as well as a separate object detailing chances of each unique individually. Will need to be changed to account for pet drop as well as linear interpolation after the summer sweepup changes.