Built flask integration, added basic front end
Result currently displays table in example as follows:
{
  "inputs": {
    "party_points": 20000,
    "personal_points": 20000,
    "raid_level": 300,
    "team_size": 1
  },
  "results": {
    "party_points": 20000,
    "party_points_estimation": false,
    "party_prob": 0.044444444444444446,
    "personal_points": 20000,
    "personal_prob": 0.044444444444444446,
    "points_per_percent": 4500,
    "team_size": 1,
    "unique_table_party": {
      "Elidnis_ward": 0.005555555555555556,
      "Lightbearer": 0.012957563977972139,
      "Masori_body": 0.003703703703703704,
      "Masori_chaps": 0.003703703703703704,
      "Masori_mask": 0.003703703703703704,
      "Osmumtens_fang": 0.012957563977972139,
      "Tumekens_shadow": 0.001851851851851852
    },
    "unique_table_personal": {
      "Elidnis_ward": 0.005555555555555556,
      "Lightbearer": 0.012957563977972139,
      "Masori_body": 0.003703703703703704,
      "Masori_chaps": 0.003703703703703704,
      "Masori_mask": 0.003703703703703704,
      "Osmumtens_fang": 0.012957563977972139,
      "Tumekens_shadow": 0.001851851851851852
    }
  }
}

Next steps would be to clean this up. Potentially adding images with corresponding chances to acquire.

Flask app routes are fairly simple, with a base index.html and a route for the api calculation, which creates a UC object to perform calculations and make assumptions if party points not included. As of now, not much in the way of data validation on inputs. Some minor changes to formula.py to avoid errors and miscalculations.